Add command metrics to IPA API
This will allow the agent to report timing and counter data for commands
that are run on the agent (starting with Standby commands). It will allow a
properly configured agent to send metrics to a remote or locally packaged
statsd. It is also extensible, so other metrics systems can easily be used.

Change-Id: Ic08d4ff78b6fb614b474b956a32eac352a14262a
Partial-Implements: add-pluggable-metrics-backend-for-ironic-and-ipa
(cherry picked from commit d68f36f4e08160f3f4bb5e99cb8aeb141300dc05)